refactor: move pena membership slice into core application
