+++ backend/src/api/middleware/exception_mapper.py
@@ -10,7 +10,9 @@ from core.application.use_cases import (
     InvalidAdminRegistrationDataError,
     InvalidLinkTokenError,
     InvalidPenaAccountabilityDataError,
+    InvalidPenaGuestPlayerDataError,
     InvalidPenaLabelsDataError,
+    InvalidPenaMembershipUpdateDataError,
     InvalidPlayerUpdateDataError,
     InvalidRegistrationDataError,
     PenaAccessDeniedError,
@@ -20,6 +22,12 @@ from core.application.use_cases import (
     PenaAccountabilityPenaNotFoundError,
     PenaLabelsAccessDeniedError,
     PenaLabelsPenaNotFoundError,
+    PenaMembershipAccessDeniedError,
+    PenaMembershipInvalidNationalityError,
+    PenaMembershipNotFoundError,
+    PenaMembershipPenaNotFoundError,
+    PenaMembershipPlayerNotFoundError,
+    PenaMembershipUserProfileNotFoundError,
     UserAlreadyLinkedError,
     UserInvalidNationalityError,
     UserProfileNotFoundError,
@@ -42,20 +50,12 @@ from core.application.use_cases import (
 )
 from fastapi import HTTPException, status
 from persistence.application.use_cases import (
-    InvalidPenaGuestPlayerDataError,
-    InvalidPenaMembershipUpdateDataError,
     InvalidPenaProfileImageError,
     InvalidSeasonDataError,
     InvalidSeasonInsightsDataError,
     InvalidSeasonMatchDataError,
     InvalidSeasonPlayerBatchDataError,
     InvalidSeasonPlayerUpdateDataError,
-    PenaMembershipAccessDeniedError,
-    PenaMembershipInvalidNationalityError,
-    PenaMembershipNotFoundError,
-    PenaMembershipPenaNotFoundError,
-    PenaMembershipPlayerNotFoundError,
-    PenaMembershipUserProfileNotFoundError,
     PenaProfileAccessDeniedError,
     PenaProfileNotFoundError,
     SeasonMatchInvalidPlayersError,
