+++ backend/src/core/application/policies/__init__.py
@@ -1,0 +1,3 @@
+from .update_policies import FieldUpdate, StandingsUpdatePolicy, UpdatePolicy
+
+__all__ = ["FieldUpdate", "StandingsUpdatePolicy", "UpdatePolicy"]
