+++ backend/src/core/application/use_cases/__init__.py
@@ -27,6 +27,17 @@ from .manage_pena_labels_usecase import (
     PenaLabelsAccessDeniedError,
     PenaLabelsPenaNotFoundError,
 )
+from .manage_pena_membership_usecase import (
+    InvalidPenaGuestPlayerDataError,
+    InvalidPenaMembershipUpdateDataError,
+    ManagePenaMembershipUseCase,
+    PenaMembershipAccessDeniedError,
+    PenaMembershipInvalidNationalityError,
+    PenaMembershipNotFoundError,
+    PenaMembershipPenaNotFoundError,
+    PenaMembershipPlayerNotFoundError,
+    PenaMembershipUserProfileNotFoundError,
+)
 from .manage_pena_profile_usecase import (
     InvalidPenaProfileImageError,
     ManagePenaProfileUseCase,
@@ -73,7 +84,9 @@ __all__ = [
     "InvalidAdminRegistrationDataError",
     "InvalidLinkTokenError",
     "InvalidPenaAccountabilityDataError",
+    "InvalidPenaGuestPlayerDataError",
     "InvalidPenaLabelsDataError",
+    "InvalidPenaMembershipUpdateDataError",
     "InvalidPenaProfileImageError",
     "InvalidNationalityError",
     "InvalidPlayerUpdateDataError",
@@ -83,6 +96,7 @@ __all__ = [
     "LinkUserToPenaUseCase",
     "ManagePenaAccountabilityUseCase",
     "ManagePenaLabelsUseCase",
+    "ManagePenaMembershipUseCase",
     "ManagePenaProfileUseCase",
     "AdminUsernameExistsError",
     "PenaAccessDeniedError",
@@ -92,6 +106,12 @@ __all__ = [
     "PenaAccountabilityPenaNotFoundError",
     "PenaLabelsAccessDeniedError",
     "PenaLabelsPenaNotFoundError",
+    "PenaMembershipAccessDeniedError",
+    "PenaMembershipInvalidNationalityError",
+    "PenaMembershipNotFoundError",
+    "PenaMembershipPenaNotFoundError",
+    "PenaMembershipPlayerNotFoundError",
+    "PenaMembershipUserProfileNotFoundError",
     "PenaProfileAccessDeniedError",
     "PenaProfileNotFoundError",
     "PenaSeasonNotFoundError",
